+++ core/library/src/commonMain/kotlin/ro/cosminmihu/ktor/monitor/ui/detail/body/Body.kt
@@ -27,6 +27,7 @@ import ro.cosminmihu.ktor.monitor.ui.detail.DisplayMode
 import ro.cosminmihu.ktor.monitor.ui.detail.copyTextFor
 import ro.cosminmihu.ktor.monitor.ui.detail.formater.Css
 import ro.cosminmihu.ktor.monitor.ui.detail.formater.FormUrlEncoded
+import ro.cosminmihu.ktor.monitor.ui.detail.formater.JavaScript
 import ro.cosminmihu.ktor.monitor.ui.detail.formater.JsonTree
 import ro.cosminmihu.ktor.monitor.ui.detail.formater.Text
 import ro.cosminmihu.ktor.monitor.ui.detail.formater.XmlTree
@@ -93,6 +94,13 @@ internal fun Body(
             body.contentFormat == DetailUiState.ContentFormat.FORM_URLENCODED && displayMode == DisplayMode.CODE && body.raw != null ->
                 FormUrlEncoded(
                     body = body.raw,
+                    modifier = Modifier.fillMaxHeight().codeBlock(),
+                    contentPadding = PaddingValues(Dimens.Small),
+                )
+
+            body.contentFormat == DetailUiState.ContentFormat.JAVASCRIPT && displayMode == DisplayMode.CODE && body.raw != null ->
+                JavaScript(
+                    code = body.raw,
                     modifier = Modifier.fillMaxHeight().codeBlock(),
                     contentPadding = PaddingValues(Dimens.Small),
                 )