+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.12.0] - TBA
+* Context7 Integration
+* UI: client source, json viewer enhancements, lines number, display mode buttons
+
 ## [1.11.2] - 2026-04-20
 * Add RetentionPeriod and ContentLength for no-op core library